# coderd/apidoc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Coderd is the service created by running coder server. It is a thin API that connects workspaces, provisioners and users. coderd stores its state in Postgres and is the only service that communicates with Postgres.",
    "title": "Coder API",
    "termsOfService": "https://coder.com/legal/terms-of-service",
    "contact": {
      "name": "API Support",
      "url": "https://coder.com",
      "email": "support@coder.com"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://github.com/coder/coder/blob/main/LICENSE"
    },
    "version": "2.0"
  },
  "basePath": "/api/v2",
  "paths": {
...
    "/audit": {
      "get": {
        "security": [
          {
            "CoderSessionToken": []
          }
        ],
        "produces": ["application/json"],
        "tags": ["Audit"],
        "summary": "Get audit logs",
        "operationId": "get-audit-logs",
        "parameters": [
          {
            "type": "string",
            "description": "Search query",
            "name": "q",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "description": "After ID",
            "name": "after_id",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Page limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Page offset",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/codersdk.AuditLogResponse"
            }
          }
        }
      }
    },
    "/audit/testgenerate": {
      "post": {
        "security": [
          {
            "CoderSessionToken": []
          }
        ],
        "consumes": ["application/json"],
        "tags": ["Audit"],
        "summary": "Generate fake audit log",
        "operationId": "generate-fake-audit-logs",
        "parameters": [
          {
            "description": "Audit log request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/codersdk.CreateTestAuditLogRequest"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "codersdk.AuditAction": {
      "type": "string",
      "enum": ["create", "write", "delete", "start", "stop"],
      "x-enum-varnames": [
        "AuditActionCreate",
        "AuditActionWrite",
        "AuditActionDelete",
        "AuditActionStart",
        "AuditActionStop"
      ]
    },
    "codersdk.AuditDiff": {
      "type": "object",
      "additionalProperties": {
        "$ref": "#/definitions/codersdk.AuditDiffField"
      }
    },
    "codersdk.AuditDiffField": {
      "type": "object",
      "properties": {
        "new": {},
        "old": {},
        "secret": {
          "type": "boolean"
        }
      }
    },
    "codersdk.AuditLog": {
      "type": "object",
      "properties": {
        "action": {
          "$ref": "#/definitions/codersdk.AuditAction"
        },
        "additional_fields": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        },
        "description": {
          "type": "string"
        },
        "diff": {
          "$ref": "#/definitions/codersdk.AuditDiff"
        },
        "id": {
          "type": "string"
        },
        "ip": {
          "type": "string"
        },
        "is_deleted": {
          "type": "boolean"
        },
        "organization_id": {
          "type": "string"
        },
        "request_id": {
          "type": "string"
        },
        "resource_icon": {
          "type": "string"
        },
        "resource_id": {
          "type": "string"
        },
        "resource_link": {
          "type": "string"
        },
        "resource_target": {
          "description": "ResourceTarget is the name of the resource.",
          "type": "string"
        },
        "resource_type": {
          "$ref": "#/definitions/codersdk.ResourceType"
        },
        "status_code": {
          "type": "integer"
        },
        "time": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/codersdk.User"
        },
        "user_agent": {
          "type": "string"
        }
      }
    },
    "codersdk.AuditLogResponse": {
      "type": "object",
      "properties": {
        "audit_logs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/codersdk.AuditLog"
          }
        },
        "count": {
          "type": "integer"
        }
      }
    },
...
    "codersdk.CreateTestAuditLogRequest": {
      "type": "object",
      "properties": {
        "action": {
          "enum": ["create", "write", "delete", "start", "stop"],
          "allOf": [
            {
              "$ref": "#/definitions/codersdk.AuditAction"
            }
          ]
        },
        "resource_id": {
          "type": "string"
        },
        "resource_type": {
          "enum": [
            "organization",
            "template",
            "template_version",
            "user",
            "workspace",
            "workspace_build",
            "git_ssh_key",
            "api_key",
            "group"
          ],
          "allOf": [
            {
              "$ref": "#/definitions/codersdk.ResourceType"
            }
          ]
        },
        "time": {
          "type": "string"
        }
      }
    },
...
    "codersdk.ResourceType": {
      "type": "string",
      "enum": [
        "organization",
        "template",
        "template_version",
        "user",
        "workspace",
        "workspace_build",
        "git_ssh_key",
        "api_key",
        "group",
        "group_members"
      ],
      "x-enum-varnames": [
        "ResourceTypeOrganization",
        "ResourceTypeTemplate",
        "ResourceTypeTemplateVersion",
        "ResourceTypeUser",
        "ResourceTypeWorkspace",
        "ResourceTypeWorkspaceBuild",
        "ResourceTypeGitSSHKey",
        "ResourceTypeAPIKey",
        "ResourceTypeGroup",
        "ResourceTypeGroupMembers"
      ]
    },
...
    "codersdk.Role": {
      "type": "object",
      "properties": {
        "display_name": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "codersdk.User": {
      "type": "object",
      "required": ["created_at", "email", "id", "username"],
      "properties": {
        "avatar_url": {
          "type": "string",
          "format": "uri"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "last_seen_at": {
          "type": "string",
          "format": "date-time"
        },
        "organization_ids": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "uuid"
          }
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/codersdk.Role"
          }
        },
        "status": {
          "enum": ["active", "suspended"],
          "allOf": [
            {
              "$ref": "#/definitions/codersdk.UserStatus"
            }
          ]
        },
        "username": {
          "type": "string"
        }
      }
    },
    "codersdk.UserStatus": {
      "type": "string",
      "enum": ["active", "suspended"],
      "x-enum-varnames": ["UserStatusActive", "UserStatusSuspended"]
    },
...
  },
  "securityDefinitions": {
    "CoderSessionToken": {
      "type": "apiKey",
      "name": "Coder-Session-Token",
      "in": "header"
    }
  }
}